feat: update backend
updated the names in each command. These are meant for the error "The Service "server_db" is invalid: metadata.name: Invalid value: "server_db": a DNS-1035 label must consist of lower case alphanumeric characters or '-', start with an alphabetic character, and end with an alphanumeric character (e.g. 'my-name',  or 'abc-123', regex used for validation is '[a-z]([-a-z0-9]*[a-z0-9])?')".