feat: update progress
- Removed lines about RateLimiter added in previous commit.

I learned that RateLimiter inside `startup_redis` function doesn't raise error. When `kubectl logs ...` is invoked right after server setup, no errors are returned. Adding logging doesn't help, because it doesn't show on kubectl logs. It's a work of a surprise.

Adding to the surprise, I understand from previous commits that `startup_redis` function works. For instance, providing a wrong ip such as `db.redis-cluster.com` raised an error as soon as the gateway server started. If it works, the line of code shouldnt be ignored.

I wonder if there is a workaround.