feat: uncomment all CICD workflows
In previous commits, I have separately verified the successful CICD operation of 1. everthing from the beginning to the integration tests and 2. everything from the kubernetes deployment to clean up. Now, I am combining both, and see if it flawlessly executes CICD end-to-end. This update is going to complete my CICD flow.