feat: Add a solution for successful deployment
One of the requirements of passing this project is including a "Screenshot of the Circle CI or Jenkins pipeline showing all stages passed successfully". I am including this screenshot to demonstrate that my CICD pipeline is successful. With this inclusion of file, I am a step closer to submitting this project.