feat: Add a solution for successful deployment
One of the requirements of passing this project is including a "Screenshot of your AWS EC2 page showing the newly created (for blue/green) or modified (for rolling) instances running as the EKS cluster nodes". I am including this screenshot to demonstrate that my CICD pipeline is successful. With this inclusion of file, I am a step closer to submitting this project.