feat: extra make sure integration testing in CICD is working
There are times when it errors due to socket connection error (for the conversion of the kubernetes-specific domain name of a kubernetes service to IP address). According to reference #42 in README.md, this is happening because no known DNS was able to answer with the IP address. That is, pods in a service (e.g. image to jpg) are not yet ready when a request is made. This measure would extra ensure the integration testing won't fail randomly.