feat: Ensure proper cleanup during CICD failure
When a branched out step fails, but another succeeds I don't know which rollback to execute. This is a problem because a roll back command would prematurely exit and cleanup won't complete. When this happens, I need to go to AWS and manually delete the stacks. This is not good. I have reorganized the steps after integration testing as a singular pipeline, so I know which roll back commands to run. Branching out CICD pipeline isn't so good if I don't know which roll back command to run when one fails.